(PC-35117)[API] refactor: move relevant fraud check test
